--- src/main/java/nus/iss/se/magicBag/dto/Result.java
@@ -1,58 +1,0 @@
-package nus.iss.se.magicBag.dto;
-
-import lombok.Data;
-import nus.iss.se.magicBag.common.ResultEnum;
-
-
-/**
- * @author mijiupro
- */
-
-@Data
-public class Result<T> {
-
-    // 操作代码
-    Integer code;
-
-    // 提示信息
-    String message;
-
-    // 结果数据
-    T data;
-
-    public Result(ResultEnum resultCode) {
-        this.code = resultCode.getCode();
-        this.message = resultCode.getMessage();
-    }
-
-    public Result(ResultEnum resultCode, T data) {
-        this.code = resultCode.getCode();
-        this.message = resultCode.getMessage();
-        this.data = data;
-    }
-    public Result(String message) {
-        this.message = message;
-    }
-    //成功返回封装-无数据
-    public static Result<String> success() {
-        return new Result<>(ResultEnum.SUCCESS);
-    }
-    //成功返回封装-带数据
-    public static <T> Result<T> success(T data) {
-        return new Result<>(ResultEnum.SUCCESS, data);
-    }
-    //失败返回封装-使用默认提示信息
-    public static Result<String> error() {
-        return new Result<>(ResultEnum.FAIL);
-    }
-    //失败返回封装-使用返回结果枚举提示信息
-    public static Result<String> error(ResultEnum resultCode) {
-        return new Result<>(resultCode);
-    }
-    //失败返回封装-使用自定义提示信息
-    public static Result<String> error(String message) {
-        return new Result<>(message);
-
-    }
-}
-
