--- src/main/java/nus/iss/se/magicBag/exception/DecryptionFailureException.java
@@ -1,7 +1,0 @@
-package nus.iss.se.magicBag.exception;
-
-public class DecryptionFailureException extends RuntimeException {
-    public DecryptionFailureException(String message) {
-        super(message);
-    }
-}
