--- src/main/java/nus/iss/se/magicBag/security/LoginUser.java
@@ -1,50 +1,0 @@
-package nus.iss.se.magicBag.security;
-
-import nus.iss.se.magicBag.entity.User;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.io.Serializable;
-import java.util.Collection;
-import java.util.Collections;
-
-public record LoginUser(User user) implements UserDetails, Serializable {
-    //返回权限信息
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return Collections.singletonList(new SimpleGrantedAuthority("ROLE_USER"));
-    }
-
-    //返回用户密码
-    @Override
-    public String getPassword() {
-        return user.getPassword();
-    }
-
-    //返回用户账号
-    @Override
-    public String getUsername() {
-        return user.getUsername();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
