--- src/main/java/nus/iss/se/magicBag/security/UserDetailsServiceImpl.java
@@ -1,25 +1,0 @@
-package nus.iss.se.magicBag.security;
-
-import lombok.RequiredArgsConstructor;
-import nus.iss.se.magicBag.entity.User;
-import nus.iss.se.magicBag.service.IUserService;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-    private final IUserService userService;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userService.findByUsername(username);
-        if (user == null) {
-            throw new UsernameNotFoundException("user doesn't exist");
-        }
-        return new LoginUser(user);
-    }
-}
