--- src/main/java/nus/iss/se/magicBag/util/JwtUtil.java
@@ -1,71 +1,0 @@
-package nus.iss.se.magicBag.util;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import javax.crypto.SecretKey;
-import java.nio.charset.StandardCharsets;
-import java.util.Date;
-
-@Component
-public class JwtUtil {
-
-    @Value("${jwt.secret}")
-    private String secret;
-
-    @Value("${jwt.expire-time}")
-    private Long expireTime;
-
-    private SecretKey getKey() {
-        return Keys.hmacShaKeyFor(secret.getBytes(StandardCharsets.UTF_8));
-    }
-
-    /**
-     * 生成 JWT Token
-     * @param subject 主题（如用户ID）
-     * @param role 角色（user/merchant/admin）
-     * @return token
-     */
-    public String generateToken(String subject, String role) {
-        return Jwts.builder()
-                .setSubject(subject)
-                .claim("role", role)
-                .setIssuedAt(new Date())
-                .setExpiration(new Date(System.currentTimeMillis() + expireTime))
-                .signWith(getKey())
-                .compact();
-    }
-
-    /**
-     * 解析 Token 获取 Claims
-     */
-    public Claims getClaims(String token) {
-        return Jwts.parserBuilder()
-                .setSigningKey(getKey())
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-    }
-
-    /**
-     * 校验 Token 是否有效
-     */
-    public boolean isTokenValid(String token) {
-        try {
-            final String subject = getClaims(token).getSubject();
-            return !subject.isEmpty() && !isTokenExpired(token);
-        } catch (Exception e) {
-            return false;
-        }
-    }
-
-    /**
-     * 判断是否过期
-     */
-    private boolean isTokenExpired(String token) {
-        return getClaims(token).getExpiration().before(new Date());
-    }
-}
