+++ src/main/java/nus/iss/se/magicbag/dto/LoginResp.java
@@ -1,0 +1,17 @@
+package nus.iss.se.magicbag.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.Date;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class LoginResp {
+    private String username;
+    private String token;
+    private String roles;
+    private Date loginDate;
+}
