+++ src/main/java/nus/iss/se/magicbag/exception/DecryptionFailureException.java
@@ -1,0 +1,7 @@
+package nus.iss.se.magicbag.exception;
+
+public class DecryptionFailureException extends RuntimeException {
+    public DecryptionFailureException(String message) {
+        super(message);
+    }
+}
