+++ src/main/java/nus/iss/se/magicbag/service/IUserService.java
@@ -1,0 +1,11 @@
+package nus.iss.se.magicbag.service;
+
+import com.baomidou.mybatisplus.extension.service.IService;
+import nus.iss.se.magicbag.dto.LoginReq;
+import nus.iss.se.magicbag.entity.User;
+
+public interface IUserService extends IService<User> {
+    User findByUsername(String username);
+
+    boolean register(LoginReq req);
+}
