+++ src/main/java/nus/iss/se/magicbag/service/impl/UserServiceImpl.java
@@ -1,0 +1,36 @@
+package nus.iss.se.magicbag.service.impl;
+
+import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import lombok.RequiredArgsConstructor;
+import nus.iss.se.magicbag.common.ResultEnum;
+import nus.iss.se.magicbag.dto.LoginReq;
+import nus.iss.se.magicbag.entity.User;
+import nus.iss.se.magicbag.exception.UserErrorException;
+import nus.iss.se.magicbag.mapper.UserMapper;
+import nus.iss.se.magicbag.service.IUserService;
+import nus.iss.se.magicbag.util.PasswordUtil;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements IUserService{
+    private final PasswordUtil passwordUtil;
+
+    @Override
+    public User findByUsername(String username) {
+        return this.baseMapper.selectByUsername(username);
+    }
+
+    @Override
+    public boolean register(LoginReq req) {
+        User existUser = findByUsername(req.getUsername());
+        if (existUser != null){
+            throw new UserErrorException(ResultEnum.USER_HAS_EXISTED);
+        }
+
+        User user = new User();
+        user.setUsername(req.getUsername());
+        user.setPassword(passwordUtil.encode(req.getPassword()));
+        return this.save(user);
+    }
+}
