+++ src/test/java/nus/iss/se/magicbag/util/PasswordUtilTest.java
